requirements.txt
diffusers==0.25.1
openai==1.11.0
Pillow==10.2.0
python-dotenv==1.0.1
streamlit==1.30.0
torch==2.1.2
